requirements.txt
flask
flask-reuploaded
flask-login
flask-wtf
flask-restful
wtforms
flask-sqlalchemy
sqlalchemy

numpy
pandas
geopy

webbrowser
requests
beautifulsoup4
lxml
html5lib

mysql-connector-python
pillow
plotly


